requirements.txt
asgiref==3.8.1
Django==5.1.7
django-unfold==0.53.0
djangorestframework==3.15.2
jalali_core==1.0.0
jdatetime==5.2.0
pytz==2025.2
sqlparse==0.5.3
